Split out marker functions from profile-data.js
